Some minor font sizing improvements.
Also fixed a few implicit casts.

+++ es-app/src/views/SystemView.cpp
@@ -1187,8 +1187,7 @@ void SystemView::legacyApplyTheme(const std::shared_ptr<ThemeData>& theme)
                                    std::floor(mPrimary->getPosition().y) + mPrimary->getSize().y);
     mLegacySystemInfo->setBackgroundColor(0xDDDDDDD8);
     mLegacySystemInfo->setRenderBackground(true);
-    mLegacySystemInfo->setFont(
-        Font::get(static_cast<int>(0.035f * mSize.y), Font::getDefaultPath()));
+    mLegacySystemInfo->setFont(Font::get(0.035f * mSize.y, Font::getDefaultPath()));
     mLegacySystemInfo->setColor(0x000000FF);
     mLegacySystemInfo->setUppercase(true);
     mLegacySystemInfo->setZIndex(50.0f);
